terraform/modules/cloud_functions/src/page-processor/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cloudevents==1.9.0
dataclasses-json==0.5.7
deprecation==2.1.0
Flask==2.2.3
frozenlist==1.3.3
functions-framework==3.3.0
google-api-core==2.11.0
google-auth==2.17.2
google-cloud==0.34.0
google-cloud-bigquery==3.9.0
google-cloud-core==2.3.2
google-cloud-documentai==2.15.0
google-cloud-documentai-toolbox==0.4.1a0
google-cloud-pubsub==2.16.0
google-cloud-storage==2.8.0
google-cloud-vision==3.4.1
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
greenlet==2.0.2
grpc-google-iam-v1==0.12.6
grpcio==1.53.0
grpcio-status==1.53.0
gunicorn==20.1.0
idna==3.4
immutabledict==2.2.4
importlib-metadata==6.1.0
intervaltree==3.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.136
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.2
openapi-schema-pydantic==1.2.4
packaging==23.0
pandas==1.5.3
pikepdf==7.1.2
Pillow==9.5.0
proto-plus==1.22.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.7
PyPDF2==3.0.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
rsa==4.9
six==1.16.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.47
tenacity==8.2.2
tiktoken==0.3.3
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.15
watchdog==3.0.0
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.15.0
